Enable active in feeds list.
Changed $url to $display_url throughout to ensure that it remains.
Check each url in feeds list against $display_url to mark the active
one.
Changed list-group in feeds loist to be a link list.
No longer bail out completely if there is a problem reading a feed,
because that stopped the feed list being built.